Move ModulesOrder to a separate file
